moved the hitl part to llm funcs and made the deep_think_check just an example script.
